refactor(auth): move JWT and refresh logic to TokenService and RefreshTokenService
Improves separation of concerns and adheres to the Single Responsibility Principle (SRP).
